clang-tidy

+++ power_grid_model_c/power_grid_model/include/power_grid_model/common/logging.hpp
@@ -11,7 +11,7 @@
 namespace power_grid_model {
 namespace common::logging {
 
-enum class LogEvent : int32_t {
+enum class LogEvent : int16_t {
     unknown = -1,
     total = 0000,       // TODO(mgovers): find other error code?
     build_model = 1000, // TODO(mgovers): find other error code?
